add nodemon to package.json so heroku can run npm start remove authorization check on /symbols/prices fixed float formatting on stock price json response to frontend change stock price formatting to fix heroku error change stock price formatting to fix heroku error fix error in cash calculation for buy transaction had to fixed getUserCashBy() function and make it return promise prepping for 5 minute api call setInterval in place to call yahoo API database update; /users/prices now only queries the db, not the yahoo API, moment and moment-timezone installed for intervals
