+++ package.json
@@ -15,7 +15,10 @@
     "express-session": "^1.12.1",
     "jade": "~1.11.0",
     "knex": "^0.9.0",
+    "moment": "^2.11.1",
+    "moment-timezone": "^0.5.0",
     "morgan": "~1.6.1",
+    "nodemon": "^1.8.1",
     "passport": "^0.3.2",
     "passport-google-oauth": "^0.2.0",
     "pg": "^4.4.3",
